moved code to cpal module
